+++ bindings/core/src/method_handler/client.rs
@@ -189,8 +189,8 @@ pub(crate) async fn call_client_method_internal(
         ClientMethod::GetAccountCongestion { account_id, work_score } => {
             Response::Congestion(client.get_account_congestion(&account_id, work_score).await?)
         }
-        ClientMethod::GetOutputManaRewards { output_id, slot_index } => {
-            Response::ManaRewards(client.get_output_mana_rewards(&output_id, slot_index).await?)
+        ClientMethod::GetOutputManaRewards { output_id, slot } => {
+            Response::ManaRewards(client.get_output_mana_rewards(&output_id, slot).await?)
         }
         ClientMethod::GetValidators { page_size, cursor } => {
             Response::Validators(client.get_validators(page_size, cursor).await?)