+++ sdk/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 2.0.0-beta.1 - 2024-05-08
+
+### Fixed
+
+- `Client::get_output_mana_rewards()` slot query parameter;
+
 ## 2.0.0-alpha.1 - 2024-04-29
 
 Initial alpha release of the 2.0 SDK.